requirements.txt
flask~=3.1.0
flask_sqlalchemy
flask_migrate
python-dotenv~=1.0.1
Werkzeug~=3.1.3
Flask-Login~=0.6.3